Set CHROME_DESKTOP to the appropriate value
This value should always be `<app_id>.desktop`. Not setting this
is known to cause weird issues with wrong icons being displayed
by the DE (`chromium-browser` icon shown instead of the actual
browser icon under certain circumstances) and causes the icon
to not show up for a couple of seconds under GNOME Wayland
when Chromium is not configured to use Wayland with the
`--ozone-platform=wayland` flag.

Signed-off-by: rany <ranygh@riseup.net>